# research/swagger/admin_example.json
{"swagger":"2.0","info":{"description":"此示例用于(interface_research)研究用, 包含不同种类的接口形态, 研究使用swagger规范生成客户端可执行代码","version":"1.0","title":"RESTFul接口示例","contact":{"name":"yangqi"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"ExampleController","description":"Example Controller"},{"name":"UmsAdminController","description":"Ums Admin Controller"}],"paths":{"/admin/login":{"post":{"tags":["UmsAdminController"],"summary":"登录以后返回token","operationId":"loginUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"umsAdminLoginParam","description":"umsAdminLoginParam","required":true,"schema":{"$ref":"#/definitions/UmsAdminLoginParam"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CommonResultOfLoginResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"Authorization":["global"]}],"deprecated":false}},"/admin/register":{"post":{"tags":["UmsAdminController"],"summary":"用户注册","operationId":"registerUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"umsAdminParam","description":"umsAdminParam","required":true,"schema":{"$ref":"#/definitions/UmsAdminParam"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CommonResultOfUmsAdmin"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"Authorization":["global"]}],"deprecated":false}},"/admin/user_detail_info/{username}":{"get":{"tags":["UmsAdminController"],"summary":"获取用户详细信息","operationId":"getAdminDetailInfoUsingGET","produces":["*/*"],"parameters":[{"name":"username","in":"path","description":"username","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CommonResultOfUmsAdminDetailData"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"Authorization":["global"]}],"deprecated":false}},"/admin/user_info":{"get":{"tags":["UmsAdminController"],"summary":"获取用户信息","operationId":"getAdminInfoUsingGET","produces":["*/*"],"parameters":[{"name":"user_id","in":"query","description":"user_id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CommonResultOfUmsAdmin"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"Authorization":["global"]}],"deprecated":false}},"/example/get_name":{"get":{"tags":["ExampleController"],"summary":"get 无参数, 返回string","operationId":"getDataStringUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CommonResultOfstring"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"Authorization":["global"]}],"deprecated":false}},"/example/get_object_by_params":{"get":{"tags":["ExampleController"],"summary":"get 含有请求参数, 返回对象","operationId":"getObjectByParamsUsingGET","produces":["*/*"],"parameters":[{"name":"department_id","in":"query","description":"department_id","required":true,"type":"string"},{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CommonResultOfExampleObject"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"Authorization":["global"]}],"deprecated":false}},"/example/get_properties/{id}/{category}":{"get":{"tags":["ExampleController"],"summary":"get, 参数in path, 返回数组","operationId":"getPropertiesByIdUsingGET","produces":["*/*"],"parameters":[{"name":"category","in":"path","description":"类型","required":true,"type":"integer","format":"int32","x-example":3},{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CommonResultOfListOfRecordProperty"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"Authorization":["global"]}],"deprecated":false}},"/example/put_return_string_list":{"put":{"tags":["ExampleController"],"summary":"put, 返回数组","operationId":"addSomePropertiesToRecordUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"addSomeRecordParam","description":"addSomeRecordParam","required":true,"schema":{"$ref":"#/definitions/AddSomeRecordParam"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CommonResultOfListOfstring"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"Authorization":["global"]}],"deprecated":false}}},"securityDefinitions":{"Authorization":{"type":"apiKey","name":"Authorization","in":"header"}},"definitions":{"AddSomeRecordParam":{"type":"object","required":["category","id","recordProperties"],"properties":{"category":{"type":"integer","format":"int32","description":"属性分类"},"id":{"type":"string","description":"record id"},"recordProperties":{"type":"array","description":"属性列表","items":{"$ref":"#/definitions/RecordProperty"}}},"title":"AddSomeRecordParam"},"CommonResultOfExampleObject":{"type":"object","required":["code","data","message"],"properties":{"code":{"type":"integer","format":"int64","description":"状态码"},"data":{"description":"数据, 注意一般只有code=200的情况下才有该数据","$ref":"#/definitions/ExampleObject"},"message":{"type":"string","description":"提示信息, 默认为空"}},"title":"CommonResultOfExampleObject"},"CommonResultOfListOfRecordProperty":{"type":"object","required":["code","data","message"],"properties":{"code":{"type":"integer","format":"int64","description":"状态码"},"data":{"type":"array","description":"数据, 注意一般只有code=200的情况下才有该数据","items":{"$ref":"#/definitions/RecordProperty"}},"message":{"type":"string","description":"提示信息, 默认为空"}},"title":"CommonResultOfListOfRecordProperty"},"CommonResultOfListOfstring":{"type":"object","required":["code","data","message"],"properties":{"code":{"type":"integer","format":"int64","description":"状态码"},"data":{"type":"array","description":"数据, 注意一般只有code=200的情况下才有该数据","items":{"type":"string"}},"message":{"type":"string","description":"提示信息, 默认为空"}},"title":"CommonResultOfListOfstring"},"CommonResultOfLoginResponse":{"type":"object","required":["code","data","message"],"properties":{"code":{"type":"integer","format":"int64","description":"状态码"},"data":{"description":"数据, 注意一般只有code=200的情况下才有该数据","$ref":"#/definitions/LoginResponse"},"message":{"type":"string","description":"提示信息, 默认为空"}},"title":"CommonResultOfLoginResponse"},"CommonResultOfUmsAdmin":{"type":"object","required":["code","data","message"],"properties":{"code":{"type":"integer","format":"int64","description":"状态码"},"data":{"description":"数据, 注意一般只有code=200的情况下才有该数据","$ref":"#/definitions/UmsAdmin"},"message":{"type":"string","description":"提示信息, 默认为空"}},"title":"CommonResultOfUmsAdmin"},"CommonResultOfUmsAdminDetailData":{"type":"object","required":["code","data","message"],"properties":{"code":{"type":"integer","format":"int64","description":"状态码"},"data":{"description":"数据, 注意一般只有code=200的情况下才有该数据","$ref":"#/definitions/UmsAdminDetailData"},"message":{"type":"string","description":"提示信息, 默认为空"}},"title":"CommonResultOfUmsAdminDetailData"},"CommonResultOfstring":{"type":"object","required":["code","data","message"],"properties":{"code":{"type":"integer","format":"int64","description":"状态码"},"data":{"type":"string","description":"数据, 注意一般只有code=200的情况下才有该数据"},"message":{"type":"string","description":"提示信息, 默认为空"}},"title":"CommonResultOfstring"},"ExampleObject":{"type":"object","required":["id","name","type"],"properties":{"departmentId":{"type":"string"},"id":{"type":"string"},"name":{"type":"string"},"recordProperties":{"type":"array","items":{"$ref":"#/definitions/RecordProperty"}},"type":{"type":"string"}},"title":"ExampleObject"},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"tokenHead":{"type":"string"},"tokenType":{"type":"string","enum":["type1","type2"]}},"title":"LoginResponse"},"RecordProperty":{"type":"object","required":["propertyName","type","value"],"properties":{"enable":{"type":"boolean"},"propertyName":{"type":"string"},"type":{"type":"string","description":"类型, 枚举值: type1, type2, type3, 在实际业务场景中, 可以用String来代替","enum":["type1","type2","type3"]},"value":{"type":"number","format":"double"}},"title":"RecordProperty"},"UmsAdmin":{"type":"object","required":["createTime","email","icon","id","loginTime","nickName","note","password","status","username"],"properties":{"createTime":{"type":"string","format":"date-time","description":"创建时间"},"email":{"type":"string","description":"邮箱"},"icon":{"type":"string","description":"头像"},"id":{"type":"integer","format":"int64"},"loginTime":{"type":"string","format":"date-time","description":"最后登录时间"},"nickName":{"type":"string","description":"昵称"},"note":{"type":"string","description":"备注信息"},"password":{"type":"string"},"status":{"type":"integer","format":"int32","description":"帐号启用状态：0->禁用；1->启用"},"username":{"type":"string"}},"title":"UmsAdmin"},"UmsAdminDetailData":{"type":"object","required":["createTime","email","icon","id","loginTime","nickName","note","password","status","username"],"properties":{"createTime":{"type":"string","format":"date-time","description":"创建时间"},"email":{"type":"string","description":"邮箱"},"icon":{"type":"string","description":"头像"},"id":{"type":"integer","format":"int64"},"loginLogList":{"type":"array","description":"登录历史记录列表","items":{"$ref":"#/definitions/UmsAdminLoginLog"}},"loginTime":{"type":"string","format":"date-time","description":"最后登录时间"},"nickName":{"type":"string","description":"昵称"},"note":{"type":"string","description":"备注信息"},"password":{"type":"string"},"roleList":{"type":"array","description":"角色列表","items":{"$ref":"#/definitions/UmsRole"}},"status":{"type":"integer","format":"int32","description":"帐号启用状态：0->禁用；1->启用"},"username":{"type":"string"}},"title":"UmsAdminDetailData"},"UmsAdminLoginLog":{"type":"object","required":["address","adminId","createTime","id","ip","userAgent"],"properties":{"address":{"type":"string"},"adminId":{"type":"integer","format":"int64"},"createTime":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"ip":{"type":"string"},"userAgent":{"type":"string","description":"浏览器登录类型"}},"title":"UmsAdminLoginLog"},"UmsAdminLoginParam":{"type":"object","required":["password","username"],"properties":{"password":{"type":"string","description":"密码"},"username":{"type":"string","description":"用户名"}},"title":"UmsAdminLoginParam"},"UmsAdminParam":{"type":"object","required":["password","username"],"properties":{"email":{"type":"string","description":"邮箱"},"icon":{"type":"string","description":"用户头像"},"nickName":{"type":"string","description":"用户昵称"},"note":{"type":"string","description":"备注"},"password":{"type":"string","description":"密码"},"username":{"type":"string","description":"用户名"}},"title":"UmsAdminParam"},"UmsRole":{"type":"object","required":["adminCount","createTime","description","id","name","sort","status"],"properties":{"adminCount":{"type":"integer","format":"int32","description":"后台用户数量"},"createTime":{"type":"string","format":"date-time","description":"创建时间"},"description":{"type":"string","description":"描述"},"id":{"type":"integer","format":"int64"},"name":{"type":"string","description":"名称"},"sort":{"type":"integer","format":"int32"},"status":{"type":"integer","format":"int32","description":"启用状态：0->禁用；1->启用"}},"title":"UmsRole"}}}